Refactor `test_workflow.py` to use runtime test fixture.
